# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Reading Tracker",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/engagements/{engagement_id}/progress-logs": {
      "post": {
        "tags": [
          "engagements"
        ],
        "summary": "Log Progress",
        "operationId": "engagements-log_progress",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressLogCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PageProgressLogRead"
                    },
                    {
                      "$ref": "#/components/schemas/MinuteProgressLogRead"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "page": "#/components/schemas/PageProgressLogRead",
                      "minute": "#/components/schemas/MinuteProgressLogRead"
                    }
                  },
                  "title": "Response Engagements-Log Progress"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "engagements"
        ],
        "summary": "List Progress Logs",
        "operationId": "engagements-list_progress_logs",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/PageProgressLogRead"
                      },
                      {
                        "$ref": "#/components/schemas/MinuteProgressLogRead"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "type",
                      "mapping": {
                        "page": "#/components/schemas/PageProgressLogRead",
                        "minute": "#/components/schemas/MinuteProgressLogRead"
                      }
                    }
                  },
                  "title": "Response Engagements-List Progress Logs"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/engagements/{engagement_id}/progress-logs/{log_id}": {
      "patch": {
        "tags": [
          "engagements"
        ],
        "summary": "Update Progress Log",
        "operationId": "engagements-update_progress_log",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          },
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Log Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProgressLogUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/PageProgressLogRead"
                    },
                    {
                      "$ref": "#/components/schemas/MinuteProgressLogRead"
                    }
                  ],
                  "discriminator": {
                    "propertyName": "type",
                    "mapping": {
                      "page": "#/components/schemas/PageProgressLogRead",
                      "minute": "#/components/schemas/MinuteProgressLogRead"
                    }
                  },
                  "title": "Response Engagements-Update Progress Log"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "engagements"
        ],
        "summary": "Delete Progress Log",
        "operationId": "engagements-delete_progress_log",
        "parameters": [
          {
            "name": "engagement_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Engagement Id"
            }
          },
          {
            "name": "log_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Log Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MinuteProgressLogRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "engagement_id": {
            "type": "string",
            "format": "uuid",
            "title": "Engagement Id"
          },
          "logged_on": {
            "type": "string",
            "format": "date",
            "title": "Logged On"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "new_ground": {
            "type": "boolean",
            "title": "New Ground"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "type": {
            "type": "string",
            "const": "minute",
            "title": "Type",
            "default": "minute"
          },
          "minute_start": {
            "type": "integer",
            "title": "Minute Start"
          },
          "minute_end": {
            "type": "integer",
            "title": "Minute End"
          }
        },
        "type": "object",
        "required": [
          "id",
          "engagement_id",
          "logged_on",
          "created_at",
          "new_ground",
          "note",
          "minute_start",
          "minute_end"
        ],
        "title": "MinuteProgressLogRead"
      },
      "PageProgressLogRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "engagement_id": {
            "type": "string",
            "format": "uuid",
            "title": "Engagement Id"
          },
          "logged_on": {
            "type": "string",
            "format": "date",
            "title": "Logged On"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "new_ground": {
            "type": "boolean",
            "title": "New Ground"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "type": {
            "type": "string",
            "const": "page",
            "title": "Type",
            "default": "page"
          },
          "page_start": {
            "type": "integer",
            "title": "Page Start"
          },
          "page_end": {
            "type": "integer",
            "title": "Page End"
          }
        },
        "type": "object",
        "required": [
          "id",
          "engagement_id",
          "logged_on",
          "created_at",
          "new_ground",
          "note",
          "page_start",
          "page_end"
        ],
        "title": "PageProgressLogRead"
      },
      "ProgressLogCreate": {
        "properties": {
          "page_start": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Start"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "minute_start": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minute Start"
          },
          "minute_end": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minute End"
          },
          "edition_length": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition Length"
          },
          "logged_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logged On"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ProgressLogCreate"
      },
      "ProgressLogUpdate": {
        "properties": {
          "logged_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logged On"
          },
          "page_end": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Page End"
          },
          "minute_end": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Minute End"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "type": "object",
        "title": "ProgressLogUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}